Improved definition of Known Bluffs
The previous definition of Known Bluffs was a type of bluff that didn't say what cards could be used for it, however specified situations where it was to be applied. In reality a 3 Bluff where both matching 2's are visible is still a 3 Bluff, but a Known 3 Bluff, because the rule that allows the bluff to be valid is the 3 Bluff convention, the Known part simply specifies why the player isn't cluing the 2's with a bad clue.