Update Reference.md (#267)
This convention confuses me, because usually a prompt would have very high priority in lategame, and could be seen as a way to change the priorities of the player. Also, in the given example, blue clue to the blue 5 touches trash on either side of the 5 and makes no sense. I think what's intended is a play clue to red 5, which would indeed be useless, and hence trigger a finesse to move the red 4 into a different hand.