Add another example of Good Touch Principle (#369)
* Add third example of good touch principle

This was made by hacking together pieces of critical_save.png. It's not the prettiest, but it works.

* Add another example of Good Touch Principle

This was motivated by a game that I recently played where I didn't pick up on a slightly more nuanced application of GTP. Considering that GTP is so important, it's better to enumerate this case out specifically to really drive the point home, especially in a guide for beginners.