Update level-14.md
I deleted the hard 5 double bluff, because it is at odds with our current low score phase treatment of 5s; this example is treated as a 5 pull for r4.
I would also like to add some or all of the following three instructive examples, but do not do so for lack of my ability to create diagrams.

The first is the hard 5 double bluff, with r1b2g3y3p3 on the stacks, and r2 r3 blindplaying. I'd like to emphasize that the 5 can be both red or blue.
The second is a double bluff with number that initially looks like a finesse, such as cluing a non-matching 2 to get two non-matching 1s to play.
The third is a double bluff with color in a rainbow variant, where the first card played matches. For example, if it is done with green at the start of the game and g1p1 play, the card should be noted as g3g4m2m3m4.

The point of each of these examples is to illustrate how Donald has to note his card, since truly that is the only part of the double bluff that isn't just automatic.

I also renamed the existing double bluff to double total bluff for distinction.