Add multiple fresh 1's example
The docs say that 1's in the starting hand are special and play oldest
first. This implies that fresh 1's are normal and play newest first.
However, this is not very explicit, and some players get confused by
this.

Add an example to call out that, with multiple fresh 1's, they play
newest first.