Update level-9.mdx
The word "Locked" is defined as having a fully clued hand. Thus, if you do NOT have a locked hand, and you happen to be at 8 clues and can't discard, and you save a chop card, it would be weird to call that a "locked hand save", because you do not have a locked hand.

Saving a card on chop when at 8 clues is legal as 8CS and should be called such instead.