Add colours to the yes/no nodes for clue focus (#1290)
* Add colours to the yes/no nodes for clue focus

This commit demonstrates how to set node background colours from outside the flowchart definitions, with minimal extra code needed within the flowchart itself.

Inside the flowchart, we need a line for each CSS class we want to add to nodes within the chart. In the global CSS file, we define colours as variables and apply those colours.

This is intended as a proof-of-concept, rather than as a final version. I have not put much thought into whether the colours are good; I merely wanted them to be clearly present and based on the user's light/dark preference.

Signed-off-by: 3geek14 <nerd.of.pi@gmail.com>

* Update custom.css

---------

Signed-off-by: 3geek14 <nerd.of.pi@gmail.com>
Co-authored-by: James <5511220+Zamiell@users.noreply.github.com>