Reword out-of-order condition on tempo clues (#1310)
* Reword out-of-order condition on tempo clues

With the old wording, I interpreted "out of order" to only apply to cards in the same suit, where one of them _must_ play before the other _can_ play. With multiple unknown 2s, there's no order in which the cards are required to play, but the left one can still block a prompt on the right one, preventing a playable 3 from being clued.

* Update level-6.mdx

* Update level-6.mdx

---------

Co-authored-by: James <5511220+Zamiell@users.noreply.github.com>