Clarifying Exception to Extinguishing Every Clue in Early Game
I was playing with a beginner who had probably read ahead but was still playing at lvl 1 and ended early game when a 2 save was available to him. He explained in the review that since that 2 save would have bad-touched another 2 that was on the stacks he thought he was exempt from cluing it. He said he had read this in the doc but we couldn't find it in the level 1 part of the doc. This is the only place where I've been able to find that axis of confusion arising, so I'm clarifying that this exception only includes Play Clues.